requirements.txt
# Core dependencies for self-improving agent
openai>=1.0.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.13.0
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.2.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
psutil>=5.9.0
gitpython>=3.1.0
autopep8>=2.0.0
black>=23.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
